+++ spec/resources/aws_lb_spec.rb
@@ -1,6 +1,6 @@
 require_relative '../spec_helper'
 
-describe GeoEngineer::Resources::AwsAlbListenerRule do
+describe GeoEngineer::Resources::AwsLb do
   let(:alb_client) { AwsClients.alb }
 
   common_resource_tests(described_class, described_class.type_from_class_name)
@@ -16,30 +16,25 @@ describe GeoEngineer::Resources::AwsAlbListenerRule do
         }
       )
       alb_client.stub_responses(
-        :describe_listeners,
+        :describe_tags,
         {
-          listeners: [
+          tag_descriptions: [
             {
-              load_balancer_arn: "foo/bar-baz",
-              listener_arn: "listener/foo/bar-baz",
-              port: 443
+              resource_arn: "foo/bar-baz",
+              tags: [{ key: "Name", value: "foo/bar-baz" }]
             }
           ]
         }
       )
-      alb_client.stub_responses(
-        :describe_rules,
-        {
-          rules: [
-            {
-              rule_arn: "rule/foo/bar-baz",
-              priority: "69"
-            }
-          ]
-        }
-      )
-      remote_resources = GeoEngineer::Resources::AwsAlbListenerRule._fetch_remote_resources(nil)
+      remote_resources = GeoEngineer::Resources::AwsLb._fetch_remote_resources(nil)
       expect(remote_resources.length).to eq 1
+    end
+
+    it "should work if no ALB's exist" do
+      alb_client.stub_responses(:describe_load_balancers, { load_balancers: [] })
+
+      remote_resources = GeoEngineer::Resources::AwsLb._fetch_remote_resources(nil)
+      expect(remote_resources.length).to eq 0
     end
   end
 end
